requirements.txt
flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
